requirements.txt
PySide6>=6.5
PyOpenGL>=3.1.7